+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Regular pigeon that stays in the nether for some time will be converted into Charred Pigeon.
 - Added Archimedes.
 - Pigeons named "Drumstick" will have a slightly chonkier model.
+- Added Letter burning: using the item on a fire or campfire will consume the letter. Doesn't apply to Sealed Letters. 
 - Added 'has_mail' property to Mailbox block.
   - Unused in the mod, but can be useful for resourcepacks.
 - Equine Assurance Bureau notice letter sending is now triggered by taming an animal instead of at random to all players at once.